Add multicompare.sh
The new script is a copy of compare.sh with the necessary modifications
to download day data files from two different hosts via scp.

Host and user (if required) for the two hosts must be edited directly
in the script (assignments to sshhosta and sshhostb).
Both hosts should be accessible via public key auth.